[Delete #313] 불필요한 커스텀훅"useModalSocket" 삭제
